Update EntityManager to return read-only entity wrappers (#10)
This commit updates the EntityManager to return a new object
called an EntityGroup. The EntityGroup acts as a read-only wrapper
to the underlying internal reference to the entity group. This
provides a more pleasant experience when interacting with query
results, because users are only concerned with Entity objects,
but optionally can get information about their IDs (whereas
previously users main interaction point was with IDs).

Some niceties were added to the EntityGroup to make it feel
more familiar to work with, such as the Iterable interface (meaning
for...of can be used on it), .toArray(), and .forEach()